auth-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiomysql==0.2.0
redis[asyncio]==5.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.7.2
pydantic-settings==2.3.1
prometheus-client==0.20.0
python-multipart==0.0.9
sqlalchemy==2.0.30